+++ Evergreen/Private/ConvertTo-Hashtable.ps1
@@ -10,23 +10,23 @@ Function ConvertTo-Hashtable {
     [OutputType([System.Collections.Hashtable])]
     [CmdletBinding(SupportsShouldProcess = $False)]
     param (
-        [Parameter(Mandatory = $True, Position = 0, ValueFromPipeline)]
+        [Parameter(Position = 0, ValueFromPipeline)]
         $InputObject
     )
 
-    Process {
+    process {
         ## Return null if the input is null. This can happen when calling the function
         ## recursively and a property is null
-        If ($Null -eq $InputObject) {
-            Return $Null
+        if ($Null -eq $InputObject) {
+            return $Null
         }
 
         ## Check if the input is an array or collection. If so, we also need to convert
         ## those types into hash tables as well. This function will convert all child
         ## objects into hash tables (if applicable)
-        If ($InputObject -is [System.Collections.IEnumerable] -and $InputObject -isnot [System.String]) {
+        if ($InputObject -is [System.Collections.IEnumerable] -and $InputObject -isnot [System.String]) {
             $collection = @(
-                ForEach ($object in $InputObject) {
+                foreach ($object in $InputObject) {
                     ConvertTo-Hashtable -InputObject $object
                 }
             )
@@ -34,16 +34,16 @@ Function ConvertTo-Hashtable {
             ## Return the array but don't enumerate it because the object may be pretty complex
             Write-Output -NoEnumerate -InputObject $collection
         }
-        ElseIf ($InputObject -is [PSObject]) {
+        elseIf ($InputObject -is [PSObject]) {
             ## If the object has properties that need enumeration
             ## Convert it to its own hash table and return it
             $hash = @{ }
-            ForEach ($property in $InputObject.PSObject.Properties) {
+            foreach ($property in $InputObject.PSObject.Properties) {
                 $hash[$property.Name] = ConvertTo-Hashtable -InputObject $property.Value
             }
             $hash
         }
-        Else {
+        else {
             ## If the object isn't an array, collection, or other object, it's already a hash table
             ## So just return it.
             $InputObject