+++ Evergreen/Public/Find-EvergreenApp.ps1
@@ -34,15 +34,10 @@
 
     process {
         if ($PSBoundParameters.ContainsKey('Name')) {
-            try {
-                # If the -Name parameter is specified, filter the included manifests for that application
-                Write-Verbose -Message "Filter for: $Name."
-                $Manifests = $Manifests | Where-Object { $_.Name -match $Name }
-            }
-            catch {
-                throw $_
-            }
-            if ($Null -eq $Manifests) {
+            # If the -Name parameter is specified, filter the included manifests for that application
+            Write-Verbose -Message "Filter for: $Name."
+            $Manifests = $Manifests | Where-Object { $_.Name -match $Name }
+            if ($null -eq $Manifests) {
                 Write-Warning -Message "Omit the -Name parameter to return the full list of supported applications."
                 Write-Warning -Message "Documentation on how to contribute a new application to the Evergreen project can be found at: $($script:resourceStrings.Uri.Docs)."
                 throw "Cannot find application: $Name."
@@ -50,31 +45,22 @@
         }
 
         #region Output details from the manifest/s
-        if ($Manifests.Count -gt 0) {
-            foreach ($manifest in $Manifests) {
-                try {
-                    # Read the JSON manifest and convert to an object
-                    $Json = Get-Content -Path $manifest.FullName | ConvertFrom-Json
-                }
-                catch {
-                    throw $_
-                }
-
-                if ($Null -ne $Json) {
-                    # Build an object from the manifest details and file name and output to the pipeline
-                    $PSObject = [PSCustomObject] @{
-                        Name        = [System.IO.Path]::GetFileNameWithoutExtension($manifest.Name)
-                        Application = $Json.Name
-                        Link        = $Json.Source
-                    }
-                    Write-Output -InputObject $PSObject
-                }
+        foreach ($manifest in $Manifests) {
+            try {
+                # Read the JSON manifest and convert to an object
+                $Json = Get-Content -Path $manifest.FullName | ConvertFrom-Json
             }
-        }
-        else {
-            Write-Warning -Message "Omit the -Name parameter to return the full list of supported applications."
-            Write-Warning -Message "Documentation on how to contribute a new application to the Evergreen project can be found at: $($script:resourceStrings.Uri.Docs)."
-            throw "Failed to return application manifest."
+            catch {
+                throw $_
+            }
+
+            # Build an object from the manifest details and file name and output to the pipeline
+            $PSObject = [PSCustomObject] @{
+                Name        = [System.IO.Path]::GetFileNameWithoutExtension($manifest.Name)
+                Application = $Json.Name
+                Link        = $Json.Source
+            }
+            Write-Output -InputObject $PSObject
         }
         #endregion
     }