+++ tests/Private/ConvertTo-Hashtable.Tests.ps1
@@ -3,7 +3,7 @@
         Private Pester function tests.
 #>
 [OutputType()]
-[Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseDeclaredVarsMoreThanAssignments", "", Justification="This OK for the tests files.")]
+[Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseDeclaredVarsMoreThanAssignments", "", Justification = "This OK for the tests files.")]
 param ()
 
 BeforeDiscovery {
@@ -15,10 +15,16 @@ BeforeAll {
 Describe -Name "ConvertTo-Hashtable" {
     Context "Test conversion to hashtable" {
         It "Converts a PSObject into a hashtable" {
-            InModuleScope Evergreen {
+            InModuleScope -ModuleName "Evergreen" {
                 $ps = [PSCustomObject]@{ Name = "Name1"; Address = "Address1" }
                 $object = $ps | ConvertTo-Hashtable
                 $object | Should -BeOfType "Hashtable"
+            }
+        }
+
+        It "Returns Null if sent Null" {
+            InModuleScope -ModuleName "Evergreen" {
+                ConvertTo-Hashtable | Should -BeNullOrEmpty
             }
         }
     }